\name{dims}
\Rdversion{1.1}
\alias{dims}
\title{ dims }
\description{
The dims function returns a vector of current Processing canvas size.
}
\usage{
dims()
}
\arguments{None.}
\value{A two-element vector with the canvase width, height.}
\references{http://www.processing.org }
\author{ B. W. Lewis <bwaynelewis@gmail.com}
\examples{
# dims ()
}
