\name{mouse}
\Rdversion{1.1}
\alias{mouse}
\title{ mouse }
\description{
The mouse function returns a vector of current Processing mouse coordinates.
}
\usage{
mouse()
}
\arguments{None.}
\value{A two-element vector with the current mouse coordinates.}
\references{http://www.processing.org }
\author{ B. W. Lewis <bwaynelewis@gmail.com}
\examples{
# mouse ()
}
