\name{R4P-package}
\Rdversion{1.1}
\alias{R4P-package}
\alias{R4P}
\docType{package}
\title{R4P }
\description{ A simple R library for processing. }
\details{
\tabular{ll}{
Package: \tab R4P\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2009-08-31\cr
License: \tab GPL\cr
LazyLoad: \tab yes\cr
}
R4P provides a simple interface between R and Processing, allowing Processing programs to make R function calls and vice versa.
}
\author{
Michael Kane and B. W. Lewis

Maintainer: B. W. Lewis <bwaynelewis@gmail.com>
}
\references{http://www.processing.org, http://www.r-project.org }
\keyword{ package }
